chore: assign new issues to project board
This adds a new workflow that will assign any added issues to the Boost project - https://github.com/orgs/filecoin-project/projects/29. This is a sibling to https://github.com/filecoin-project/boost/pull/423

I've added a BOOST_BOARD secret to the boost and boost-docs repo with minimal permissions to perform the action of assigning issues to the project via the graphql endpoint. This token is set to a 90 day expiration and will need to be rotated by then.